+++ back_end/view.py
@@ -1,5 +1,5 @@
 from auth import requires_login
-from backend.common import *
+from back_end.common import *
 from template_engine import render
 from os import path
 from db import db_api as db
